Create Contact Bug

+++ models/partner_inherit.py
@@ -32,14 +32,15 @@ class PartnerInherit(models.Model):
     def _onchange_bd_tag_ids(self):
         _logger.error("called _onchange_bd_tag_ids")
         _logger.error(str(self.id) + " has " + str(len(self.bd_tag_user_ids)))
-        self._cr.execute('delete from contact_team_users where contact_id = %s', [self._origin.id])
-        _logger.error(str(self.id) + " has " + str(len(self.bd_tag_user_ids)))
+        if(self._origin.id):
+            self._cr.execute('delete from contact_team_users where contact_id = %s', [self._origin.id])
+            _logger.error(str(self.id) + " has " + str(len(self.bd_tag_user_ids)))
 
-        for bd_tag in self.bd_tag_ids:
-            users = self.env['res.users'].sudo().search(['|', ('sale_team_id.name', 'ilike', bd_tag.name), ('groups_id.name','=','User: All Documents')])
-            _logger.error("For tag name " + str(bd_tag.name) + " " + str(len(users)) + " type " + str(type(self.bd_tag_user_ids)))
-            for user in users:
-                self._cr.execute('insert into contact_team_users (user_name, user_id, contact_id) values(%s, %s, %s)', ( user.name, user.id ,self._origin.id))
+            for bd_tag in self.bd_tag_ids:
+                users = self.env['res.users'].sudo().search(['|', ('sale_team_id.name', 'ilike', bd_tag.name), ('groups_id.name','=','User: All Documents')])
+                _logger.error("For tag name " + str(bd_tag.name) + " " + str(len(users)) + " type " + str(type(self.bd_tag_user_ids)))
+                for user in users:
+                    self._cr.execute('insert into contact_team_users (user_name, user_id, contact_id) values(%s, %s, %s)', ( user.name, user.id ,self._origin.id))
 
     @api.onchange('user_id')
     def _onchange_salesperson(self):
